factored out utils
